Migrate MOE to Dagger2
First of several steps to refactor the internals.

1. Dagger 1 -> Dagger 2 (achieves a core library goal of eliminating D1)
2. Add a directive/task scope
    - provides a home within the injected graph for other parts.
3. Migrate any remaining code that uses the static to be created by dagger or autofactory
4. Remove the static AppContext.RUN (now Injector.INSTANCE)
5. Re-work the google-centric stuff into a map-bound hook, so it can be simply omitted
6. Push the re-organized code out to open-source.
6a. Migrate java moe public source to github
...
N. Profit.

In the process the tests should get a LOT simpler as they are so boiler-plate heavy because of the need to satisfy the static dependency on RUN.  As more objects can be configured by passed-in dependencies, test-setup will shrink.
Tested:
  [] test [][]
  Execute the built client:Moe binary against the dagger project.
	Change on 2015/04/17 by cgruber <cgruber@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91460335