Rework Repository/RepositoryService and ProjectContext creation in preparation for making it all injectable via a multibinder
Some notes.  RepositoryService.Factory got a type() to permit us to interrogate the factory for its type (svn, git, etc.).  This is done to remove reliance on RepositoryType which is an enum, but which is therefore brittle to different types of repositories, and I want the types to be pluggable so they're more amenable to use (and extension) in multibinders.

Some things here are begging for AutoFactory and will get that treatment, but not in this commit.

Svn got more of a cleanup, as it was a bit of a test of how the repo classes would be structured.  I didn't do the full extraction of Injector.INSTANCE from the others, because I want to do that in a separate step, and I might do things a bit differently from how I did Svn (and I'll back-port Svn into that approach then).

This CL is WAY bigger than I'd like, but things are pretty tangled, so it was hard to keep this one small.
	Change on 2015/05/18 by cgruber <cgruber@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93893279