Automated g4 rollback of changelist 93893279.
*** Reason for rollback ***

Broke several projects' open-source tests.

*** Original change description ***

Rework Repository/RepositoryService and ProjectContext creation in preparation for making it all injectable via a multibinder

Some notes.  RepositoryService.Factory got a type() to permit us to interrogate the factory for its type (svn, git, etc.).  This is done to remove reliance on RepositoryType which is an enum, but which is therefore brittle to different types of repositories, and I want the types to be pluggable so they're more amenable to use (and extension) in multibinders.

Some thing...

***
	Change on 2015/05/18 by cgruber <cgruber@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93908019