Add a Github API and some tests, and a directive to handle "github_pull" which converts a pull request from a github url into a staged set of changes.
The current version scans for the most recent equivalency, and pins the migration from that commit.  To the extent that the target repository and the current repo's main branches are out of sync, the runner of this directive can expect merge conflicts when rebasing/merging against the target branch after the workspace has been created.

The feature works essentially like moe magic, but performs a unidirectional merge, including any configured mapping/scrubbing steps, of the commits present in a git pull request.

This change also alters the MigrateBranchDirective to take into account equivalences, and to determine the relevant CLs by comparing the branches (as opposed to using a supplied single revision).  Any commit in the source branch which is not present in the ancestor branch is considered part of the migration.  For these purposes, the github_pull directive gets the ancestor information from the pull request metadata.

The feature uses the github API, but does not do any authentication, and is thus limited to public branches, and may be subject to github's rate-limit.
	Change on 2015/09/02 by cgruber <cgruber@google.com>
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=102173654