Bookkeeper report the successful or unsuccessful attempt to find head equivalence to the user-interface (it was swallowing success pretty solidly).
In the process, inject ProjectContext (since we now can) directly into Bookkeeper, and (where appropriate) make Bookkeeper lazily injected (so things that don't need project context can start up.

Also fixes a bug where MOE uses a linear strategy to scan commits, which is an invalid strategy from the point of view of projects which merge their moe-synced commits via a pull-request, as the linear approach would only see the merge commits, and never see the real commits.

Lastly, it re-orders the bookkeeping so head-equivalence is attempted once, first, and if it finds it, it omits bookkeeping the more expensive per-migration way.  In the new model, it only walks back migrations, noting them and checking equivalence if it can't find a head equivalence, and it short-circuits as soon as it finds such equivalence, writing only the minimal amount to the database.  For projects that have constant head-equivalence this can mean that migrations are never bothered to be written.

	Change on 2016/03/04 by cgruber <cgruber@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116383440