Rework the execution of moe in the regression tests to be in-memory (for ease of debugging)
In service of this, fix a few places where some resources were statically accessed, notably the scrubber, because it was done via relative filesystem locations, but with the change to in-memory, it's no longer being executed from a working directory, and so much be specified absolutely.  It's injected into ProjectConfigFactory and piped through the static method that builds the list of editors, but ultimately will be injected directly.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=118829545