Add an ability to add a dummy commit referencing previously registered commits by id.  This lets one create the history builder, and add commits, without having to store commits locally to build up parent histories.  The IDs must be unique, and a commit can only have a parent (or parents) already present in the history.
Note - this change doesn't make the current tests THAT much better, but is extracted from a commit in which I'm trying to test complex branching, and it's a pain to set up the histories.

Also, while I'm here, rename a test so it's referencing what it's actually testing.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=120168373