Fix original author processing to be more tolerant of bad author metadata.
The value is extracted from a field, which could be arbitrarily entered by any committer in the history.  That value can be invalid and can break `git` (at least) which is very picky about format.

MOE normalizes on git-style author metadata, so this change seeks to best-case convert to "foo <foo@blah.com>" format, and where it can't, it dumps the content into the "name" section, and adds "<undetermined_user>" as the "user" portion, issuing a warning and leaving it up to the user to fix the resulting metadata.  At least it will make it so that the change can go through.

	Change on 2016/07/29 by cgruber <cgruber@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128834341