Allow bookkeeping to be skipped using a flag during moe magic.
This should generally not be used, but can be useful where bookkeeping would require crossing a line in the commit-history where the configuration is valid on one side, but not on the other.  In this case, the magic migration process will just churn without attempting to pre-guess the equivalence state, using the last known, so at least the magic run can get up to the point of the change.  Then equivalence can be manually marked, the config changed, and the remainder synced according to the new world.

Also, fix a bug where provision happened during --help and so missing flags (which aren't needed for help) blew up.

	Change on 2016/07/29 by cgruber <cgruber@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128861728