Excise FileSystem from Codebase and make it an AutoValue.
Also pull a few direct references to Injector.INSTANCE into fields so they're more structurally how they would be when they're injected, and leave TODOs.

A few places I swapped in Mockito, but the goal here is not a transition - it was just where it was really hard to fix without rewriting the test, and it was easier to rewrite in Mockito.

	Change on 2016/08/22 by cgruber <cgruber@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=131000961