Rework CodebaseMerger based on some suggestions of ronshapiro@ to make it more injectable, and cleaner overall.
Essentially move the non-global things into the callstack, and let it be a more injectable object.

Also, migrate the CodebaseMergerTest to mockito, since the EasyMock interaction stuff was oppressive. The Mockito stuff isn't much better, as this is a mock-heavy test, but that avoids filesystem interaction, and keeps this in the realm of a unit test, so I left it mock-heavy.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188124751