Extract the field parsing logic, and move it into the RevisionHistory type tree, since with the exception of tests, it really only should happen when RevisionMetadata is first created, not every time it's manipulated by a scrubber.  This opens up the ability to make per-repo-type field parsing, so that piper fields and git trailers etc. can be treated differently.
Currently, it's all passed back through legacy field parsing logic, but by way of RevisionHistory.  Each type can be updated independently now.  Also, writing out the appropriate field type is not in this CL, and will have to come in another CL, as the Migrator will need to be instrumented to know about the repo-type so it can get un-parse logic for that repo.

Part of a series of commits intended to make metadata field processing more supple, so as to properly output well known fields such as are consumed by Google's CLAbot on github.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191962238