Enable kotlin in the MOE codebase expressions package, using data classes in place of @AutoValue.
In the process, merge AbstractExpression with Expression, as AE is a needless layer.  Also, remove BinaryExpression which was needed for the validation logic, but after some consideration, that validation logic is overzealous, and all the Expression creation is basically local to the expressions package (except for RepositoryExpression which is the base case anyway). The whole validation bit was just overengineered.

Note: This definitely breaks the opensource, despite my attempts to make it work.  I've definitely updated the maven stuff, and it ought to work, but it's got its issues. I'm finding it slow going, fixing it by rebuilding MOE stuff, so I'm going to take this as a two-pronged approach, namely get this in and working internally, and then fixing the external build in a separate commit (possibly by fixing hte maven bits, or possibly my migrating to bazel).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=208061385